Hotfix for 404 spamming the logging system
Not really the way we want to do it, but it works. Can't figure out what
Exceptions get thrown on aws, even after searching for a few hours I did
not find a solution. This hotfixes the issue by disabling logging for
the static requests all together (which should not be an issue), because
of time reasons only this hotfix is applied. Perhaps a better solution
will come later.